moved fz_new_isolated_smask_device to wrap-mupdf.c
